+++ src/main/java/service/Dispatcher.java
@@ -1,0 +1,4 @@
+package service;
+
+public class Dispatcher {
+}
